adds fetch requests component
