+++ functions/.gitignore
@@ -1,1 +1,2 @@
 node_modules/
+portatoapp-firebase*
