+++ src/Components/FetchRequests.tsx
@@ -1,11 +1,9 @@
 import React, { useEffect, useState } from 'react';
-import { getAuth } from 'firebase/auth';
 import { fetchDataOnce } from '../linksStoreToFirebase';
-import styles from './UserRequests.module.css';
 import { IObjectInfo } from '../type';
-import { Card, Typography, Image } from 'antd';
+import { Card } from 'antd';
 
-const UserRequests: React.FC = () => {
+const FetchRequests: React.FC<{ uid: string }> = ({ uid = undefined }) => {
   const [requests, setRequest] = useState<IObjectInfo[]>([]);
 
   // Mehdi : Use Effect to only fetch the data once when the component is mount
@@ -17,17 +15,12 @@ const UserRequests: React.FC = () => {
       // DO nothing
     });
 
-    const auth = getAuth();
-    const currentUser = auth.currentUser;
-
-    if (currentUser) {
-      const uid = currentUser.uid;
-      const data = fetchDataOnce(uid);
-      if (data) {
-        user_requests = data.then((storesArray) => {
-          return storesArray;
-        });
-      }
+    // TODO: Handle uid undefined case
+    const data = fetchDataOnce(uid);
+    if (data) {
+      user_requests = data.then((storesArray) => {
+        return storesArray;
+      });
     }
 
     const getUserRequests = async () => {
@@ -41,17 +34,6 @@ const UserRequests: React.FC = () => {
   console.log(containerHeight + 'px');
   return (
     <div>
-      <h1
-        style={{
-          marginTop: '10vh',
-          display: 'flex',
-          flexDirection: 'column',
-          justifyContent: 'center',
-          alignItems: 'center',
-        }}
-      >
-        Your Current Requests
-      </h1>
       <div style={{ height: containerHeight + 'px', overflowY: 'scroll' }}>
         {requests.map((request) => (
           <div
@@ -76,4 +58,4 @@ const UserRequests: React.FC = () => {
   );
 };
 
-export default UserRequests;
+export default FetchRequests;
